pyproject.toml
[project]
name = "data2026"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
	"python-pptx>=1.0.2",
	"streamlit>=1.44.0",
]
